App componentの中身をHome componentに移動する
