+++ front/src/App.tsx
@@ -1,74 +1,17 @@
-import React from 'react';
+import React, { FC } from 'react';
 import './App.css';
-import {
-  Button,
-  Container,
-  Header,
-  Icon,
-  Segment,
-  Menu,
-  List,
-} from 'semantic-ui-react';
-import StroyList, { Story } from './containers/StoryList';
+import { Switch, Route, Redirect } from 'react-router-dom';
+import Home from './components/Home';
 import About from './components/About';
 
-const App: React.FC = () => {
-  const a = Array.from(Array(13).keys());
-  const stories = a.map(
-    (n: number): Story => ({
-      id: n,
-      header: 'Matthew',
-      meta: 'author',
-      description: 'Matthew is a musician living in Nashville.',
-    }),
-  );
-
-  return (
-    <div className="container">
-      <main>
-        <Segment
-          inverted
-          textAlign="center"
-          style={{ minHeight: 400, padding: '1em 0em' }}
-          vertical
-        >
-          <Menu size="large">
-            <Container>
-              <Menu.Item as="a" active>
-                Home
-              </Menu.Item>
-              <Menu.Item position="right">
-                <Button as="a">Log in</Button>
-              </Menu.Item>
-            </Container>
-          </Menu>
-          <Container text>
-            <Header as="h1" inverted content="back story" />
-            <Button primary size="huge">
-              ストーリーを書いてみる
-              <Icon name="arrow right" />
-            </Button>
-          </Container>
-        </Segment>
-
-        <StroyList stories={stories} />
-
-        <Segment inverted vertical style={{ padding: '5em 0em' }}>
-          <Container>
-            <Header inverted as="h4" content="About" />
-            <List link inverted>
-              <List.Item as="a">bstとは</List.Item>
-              <List.Item as="a">作者</List.Item>
-              <List.Item as="a">問い合わせ</List.Item>
-            </List>
-            <Switch>
-              <Route path="/abount" component={About} />
-            </Switch>
-          </Container>
-        </Segment>
-      </main>
-    </div>
-  );
-};
+const App: FC<{}> = () => (
+  <div className="container">
+    <Switch>
+      <Route path="/about" component={About} />
+      <Route path="/" component={Home} />
+      <Redirect to="/" />;
+    </Switch>
+  </div>
+);
 
 export default App;
